models/11-torch_models/requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
cycler==0.11.0
fonttools==4.38.0
idna==3.4
kiwisolver==1.4.4
matplotlib==3.5.3
numpy==1.21.6
packaging==23.0
Pillow==9.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.28.2
six==1.16.0
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
typing_extensions==4.5.0
urllib3==1.26.14
torchsummary
tensorboard
torchmetrics